docs(indexes): update code sample
