--- src/indexes/_example/display.json
@@ -1,4 +1,0 @@
-{
-  "navigationBarTitleText": "Indexes 索引",
-  "navigationBarBackgroundColor": "#fff"
-}
