+++ src/indexes/_example/indexes.wxml
@@ -4,10 +4,10 @@
   <t-demo title="01 类型" desc="基础索引类型">
     <view class="box">
       <t-button t-class="type-button" variant="plain" size="large" block bind:tap="tapShowLetter">
-        <navigator url="/pages/indexes/display">字母索引</navigator>
+        <navigator url="/pages/indexes/display/index">字母索引</navigator>
       </t-button>
       <t-button t-class="type-button" variant="plain" size="large" block bind:tap="tapShowNumber">
-        <navigator url="/pages/indexes/display">数字索引</navigator>
+        <navigator url="/pages/indexes/display/index">数字索引</navigator>
       </t-button>
     </view>
   </t-demo>
